Rename main.py to endpoints.py; rename start.py to main.py
